+++ ipv6-rules.sh
@@ -45,8 +45,10 @@ function __fill_trustlist() {
     echo "2001:638:a000:4140::ffff:189 2001:678:558:1000::244 2001:67c:289c::9 2001:858:2:2:aabb:0:563b:1526 2610:1c0:0:5::131 2620:13:4000:6000::1000:118"
 
     getent ahostsv6 snowflake-01.torproject.net. snowflake-02.torproject.net. | awk '{ print $1 }'
-    curl -s 'https://onionoo.torproject.org/summary?search=flag:authority' -o - |
-      jq -cr '.relays[] | .a | select(length > 1) | .[1]' | grep ':' | tr -d ']['
+    if relays=$(curl -s 'https://onionoo.torproject.org/summary?search=flag:authority' -o -); then
+      jq -cr '.relays[]  | .a | select(length > 1) | .[1]' <<<$relays |
+        grep ':' | tr -d ']['
+    fi
   ) |
     xargs -r -n 1 -P $jobs ipset add -exist $trustlist
 }
@@ -56,11 +58,13 @@ function __fill_multilist() {
     if [[ -s /var/tmp/$multilist ]]; then
       cat /var/tmp/$multilist
     fi
-    curl -s 'https://onionoo.torproject.org/summary?search=type:relay' -o - |
-      jq -cr '.relays[] | .a | select(length > 1) | .[1]' | grep ':' | tr -d '][' | sort | uniq -d |
-      tee /var/tmp/$multilist.new
-    if [[ -s /var/tmp/$multilist.new ]]; then
-      mv /var/tmp/$multilist.new /var/tmp/$multilist
+    if relays=$(curl -s 'https://onionoo.torproject.org/summary?search=type:relay' -o -); then
+      jq -cr '.relays[] | .a | select(length > 1) | .[1]' <<<$relays |
+        grep ':' | tr -d '][' | sort | uniq -d |
+        tee /var/tmp/$multilist.new
+      if [[ -s /var/tmp/$multilist.new ]]; then
+        mv /var/tmp/$multilist.new /var/tmp/$multilist
+      fi
     fi
   ) |
     xargs -r -n 1 -P $jobs ipset add -exist $multilist
